requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-yasg==1.20.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1